添加Mapper和SQLBuilder文件夹结构，添加用户数据导出的相关数据层逻辑
